Optimize DeltaHashTable family: skipLSlot, single-pass put, hash elision, O(1) merge
Four performance optimizations to the core indexing data structure, plus
comprehensive test hardening (25 new test cases).

1. skipLSlot() — zero-allocation bitOffset() [HIGH impact]

   LSlotCodec::bitOffset() previously called decode() for each preceding
   lslot, which allocated vectors (packed_versions, ids, entries) on every
   call only to throw them away. On the hottest read path (findFirst,
   findAll), with 32 lslots per bucket the average skip count is 16 —
   meaning 16 full decode-and-discard cycles per lookup.

   New skipLSlot() reads the same unary/gamma bit markers to advance the
   BitReader position without any heap allocation. bitOffset() now loops
   over skipLSlot() instead of decode().

2. Eliminate double-lookup in GlobalIndex::put / SegmentIndex::put [HIGH]

   Both called contains() then addEntry(), performing two independent
   hash → decode → chain-walk cycles. GlobalIndex::put additionally
   acquired the bucket spinlock twice.

   addToChain() now returns bool indicating whether the fingerprint group
   was newly created, tracking an is_new flag across the entire overflow
   chain. New addEntryIsNew() on both ReadOnlyDeltaHashTable and
   ReadWriteDeltaHashTable propagates this return value. GlobalIndex::put
   and SegmentIndex::put replaced with a single addEntryIsNew() call.

3. Avoid redundant hashing in ReadWriteDeltaHashTable [MEDIUM impact]

   Locked wrappers (findAll, findFirst) hashed the key to find the bucket
   lock, then the base class method hashed the key again internally.

   New protected findAllByHash(bi, li, fp, ...) and findFirstByHash(bi,
   li, fp, ...) accept pre-decomposed hash components. Public methods
   delegate to them. ReadWriteDeltaHashTable wrappers hash once and call
   the by-hash variants directly.

4. O(1) merge in forEachGroup [LOW impact]

   Group merging across extension chains used a linear scan over
   std::vector<Group> — O(G) per entry, O(G²) total. Replaced with
   std::unordered_map<uint64_t, Group> keyed on (lslot << fp_bits) |
   fingerprint for O(1) amortized lookup.

New tests:
  - skipLSlot correctness (SkipLSlotMatchesDecode, SkipEmptyLSlots,
    SkipLSlotLargePayload)
  - addEntryIsNew semantics (first add, duplicate, different keys,
    overflow, keyCount verification)
  - findFirst across overflow chains (exact max, descending insert order,
    single-entry-per-bucket)
  - findAll completeness across overflow chains
  - forEachGroup merge correctness with overflow
  - ReadWriteDHT concurrent operations (concurrent add+findFirst,
    different keys, overflow same bucket, concurrent add+contains,
    findFirst during concurrent add)
  - Edge cases (empty table findFirst/findAll, single entry, max
    packed_version)

+++ src/db.cpp
@@ -118,7 +118,7 @@ Status DB::open(const std::string& path, const Options& options) {
         for (uint32_t id : segment_ids) {
             auto* seg = storage_->getSegment(id);
             if (!seg) continue;
-            seg->index().forEach([&](uint32_t offset, uint32_t packed_ver) {
+            seg->index().forEach([&](uint32_t offset, uint64_t packed_ver) {
                 internal::LogEntry entry;
                 Status rs = seg->readEntry(offset, entry);
                 if (rs.ok()) {
